bg_remover/requirements.txt
fastapi
uvicorn
rembg
pillow
aiohttp